unittest: Improve sparse file detection for ZFS
We now use SEEK_HOLE to detect sparse file support and allow for
a block size of up to 128K.

The existing check didn't work well on ZFS where st_blocks reports
the number of compressed blocks, and also lags behind when data is
is committed to the journal but not yet written to the file itself.
This means we can't use st_blocks here because it could suggest a
sparse file when there isn't actually support.

SEEK_HOLE is specified by POSIX and seems to be widely supported.
On platforms without it or on an FS with a > 128K block size we
will skip the tests involving a 4GB file, but that's acceptable.

Fixes #823, reported by someplaceguy.

(cherry picked from commit 8a05a114382916fd0ca141533e8c0e87da652280)