[glass] Protect table fds from stray writes
We only use pwrite() to write to these tables, so try to set the
file position to a really high value so that if user code or another
library (or errant code in libxapian itself) writes to our fd then
it'll fail rather than corrupting the database.

Currently this is enabled for Linux, FreeBSD, macOS and OpenBSD.

Partly addresses #651.