Fix stats requested by DiceCoeffWeight
Drop DOC_LENGTH_MIN which is not used (it was by an early version of the
code during development).

Explicitly request WQF which is used (not requesting it is currently
harmless as get_wqf() always works anyway but we should correctly
request the stats used in case that changes in the future.