Make stat_flags constants more optimisable
Use 0 for stats that are currently available by default, and the
same value for TERMFREQ, RELTERMFREQ and COLLECTION_FREQ which
are currently fetched together if any of them are used.  This
means the compiler can optimise the need_stat() calls down to
a simple numeric constant in more cases.  The trade-off is we
have less scope for refactoring stat handling during a release
series, but in practice that hasn't been something we've really
wanted to do, and it can just wait for the next release series.