Fix DiceCoeffWeight::get_sumpart() when wdf == 0
Remove the special case to return a weight of 0 in this case, as
it's just wrong.

Currently for an actual term the wdf is supplied even if not
requested so to trigger this requires an OP_SYNONYM query, as in
that case we are lazy about calculating wdf when it hasn't been
requested.