Support whole-db bounds in weight stats
The existing bounds are for the shard, which is usually what we want
want as with a sharded database it gives tighter bounds and so better
match optimisation, but it does make them unsuitable for uses such as
calculating a suitable offset to add to every get_sumextra() if the
weighting formula can give a negative term independent weight
contribution.

Add new bounds DB_DOC_LENGTH_MIN, DB_DOC_LENGTH_MAX,
DB_UNIQUE_TERMS_MIN, DB_UNIQUE_TERMS_MAX and DB_WDF_MAX which are for
the whole database.

Fixes #810.