Add XFAIL for checkstatsweight3 with sharded db
This testcase was previous not run for sharded databases, with a
FIXME noting this.  Investigating shows it's due to a bug where we
use the shard's termfreqs rather than those for the whole database.
I don't see a simple fix, so move it to XFAIL for now.

(cherry picked from commit b0834cbac83d2cc02faa1fe0fbb03a8e6f6684f5)