Add test coverage for $set{weighting,...}
These tests are from RELEASE/1.4 where they were added as part of
adding support for bm25+ and pl2+ there (which we get automatically
on master via Weight::create()).

These highlighted that the exception messages Weight::create() was
generating were not giving the intended context, so fix that.

They also highlighted that in 1.4 "tfidf ntn" works but we weren't
supporting these 3 character codes on master.  We ought to for
compatibility so fix that too.