Deprecate LMWeight
It was meant to implement the "Language Model" Weighting scheme, but
we've discovered the implementation was incorrect and fixing it requires
ABI-incompatible changes.

For 1.4.x we need to leave it in place so as not to break existing code.
It will be removed in the next release series and replaced with new
separate classes implementing Language Model weighting with each
smoothing.