Simplify how we clamp wdf to <= doclen
We need to do with for OP_SYNONYM where the wdf is approximated.
We now only clamp when need_stat(WDF) and need_stat(DOC_LENGTH)
have both been called - when they both have we can rely on having
both stats, so the clamping is cheap, and otherwise the weighting
scheme can't reasonably rely on wdf <= doclen.

This completely avoids the need to compute whether each wildcard or
edit distance expansion is wdf_disjoint.