[omega] Stop defining __MSVCRT_VERSION__
We fixed this for xapian-core in 1.4.24 but missed that omega defined
it too.

It looks like it's not been needed since 2015, and defining it breaks
when building with the UCRT64 variant of mingw64 (which is now the
recommended version to use).

(cherry picked from commit 26293c7076cc07dd7716cccc09b3d3ac26698a9c)