[ci] Fix showing omega build errors in VM
In the VM builds we do a quiet build first, and if that fails, repeat
the build with more verbosity so that the failing command gets shown.
This reduces the size of the logs to wade through to see the actual
problem.

However, due to a cut and paste error we were repeating the core build
if the omega build failed.