Stop explicitly linking bindings with -lstdc++
This was added in 0.9.3 in 2005 as a workaround for a linking problem
on OpenBSD, but nowadays it causes link failures on platforms which use
clang+libc++ as the C++ compiler, which includes OpenBSD so it seems
this fix is no longer useful anywhere.

If this change results in you getting a link failure, please let us
know about it.

Also try re-running make for the bindings specifying an explicit link
with -lstdc++ like so:

  make LIBS=-lstdc++

That assumes the C++ compiler you are using uses libstdc++ - if it
uses libc++ then you should probably try:

  make LIBS=-lc++

Fixes #834, reported by David Gessel.

(cherry picked from commit 3f5ea70c4375ada31e35e553146d05dc1975a6ce)