Consistently fail if --with-X can't be done
If required dependencies for bindings the user has explicitly enabled
with --with-X are missing, configure now fails in every case (previously
it would fail for some bindings but just not actually enable the
bindings for others).

The code for each binding now uses a consistent approach with a variable
such as `lua_ok` which gets set to `yes` or `no` - this helps in cases
where we want a language dependency for other purposes (since we use
Perl and Python for running scripts).