[ci] Update macos jobs
Drop the macos-12 job.  Apple have ended support for macOS 12 and
homebrew no longer supports it.  Our job still works currently but
spend ~48 minutes building all the homebrew packages it needs on
each run which seems a poor use of CPU time.

Rename the macos job to macos-14.

Add a macos-15 job.

Explicitly specify which bindings to build for each job (so we
catch if we accidentally stop detecting dependencies for one of
them) and enable more bindings.