Workaround autoconf 2.72 AC_SYS_LARGEFILE bug
This configure probe always fails when run with the C++ compiler,
like we do, which leads to large file support not being enabled
when it isn't the default (e.g. on 32-bit Linux platforms).

We really want to probe this for the C++ compiler so we now
temporarily select AC_LANG([C]) and set the C compiler to the C++
one for this one macro.

This patch can be reverted once a fixed version of autoconf is released
and we start requiring at least this fixed version.

Reported by Adrian Bunk in https://bugs.debian.org/1086249