Get kill_remote() to work in shared Microsoft Windows build
The problem is that in a shared build there's a wrapper executable
generated by libtool (like the wrapper script on Unix-like platforms)
and we were using TerminateProcess() which only kills the single
requested process.

The solution is to create a new process group for each server and
then kill all processes in that process group by sending them
CTRL_BREAK_EVENT using GenerateConsoleCtrlEvent().

The only wrinkle is that this fails under Wine (at least when using
msvcrt - not tested with ucrt), but I've failed to come up with a
reduced testcase so far.  For now we skip the test in this situation
in CI.