Improve check for broken faketime
Test the output of a small helper program which calls time() and
gmtime() and reports the year.  Previously we tested the output
of date +%Y but GNU date seems to call clock_gettime() on Linux
and that doesn't catch if faketime fails to intercept time(),
which is the case that we actually need to work.

(cherry picked from commit eeb80f62de301e3de87cc1c1cd6a2ee2325d7cf2)