Improve check for broken faketime
Test the output of a small helper program which calls time() and
gmtime() and reports the year.  Previously we tested the output
of date +%Y but GNU date seems to call clock_gettime() on Linux
and that doesn't catch if faketime fails to intercept time(),
which is the case that we actually need to work.