Improve Abiword parsing
The built-in parser can now extract metadata.

It also no longer tries to index base64 encoded data for embedded
images, etc.  Mostly this would be skipped due to being longer
than the 64 byte default word length limit, but `/` or `+` in the
base64 data, or a shorter final line could result in junk terms.