Handle fallback implementations consistently
Make them all follow the pattern where a header is included
unconditionally and the source file linked unconditionally.

The header either includes the system header which defines the function
or defines the missing function via a forwarding macro.

The source file has a #ifdef around most of the contents so
compiles to nothing on systems which provide the function.