stemtest: Add support for gzipped vocabulary lists
Until now stemtest was skipping stemdict for any language with gzip
compress vocabulary (or output), which we only have for Arabic.

Implementing this revealled that arabic/output.txt.gz was out of
date so update that to match the current version of the stemmer
we are using.  It looks like it wasn't updated to match the
changes in 36465388b7dd7d06484b02280fea693b8c71782f.