Ruby: Always link to RUBV_LIBS
We were avoiding this on platforms which allow undefined symbols in
shared libraries (which then get resolved when the Ruby interpreter
dynamically loads the module).  However this was failing to work on
macOS, and it seems the Ruby norm is to explicitly link to librubyX.Y
so we now do that.

Fixes #837, reported by Graeme Porteous.

(cherry picked from commit a3ef56cdda871cd2f6794ec50d686207a2032ae4)