[ci] Use windows-2025 instead of other images
Replace windows-2022 with windows-2025: this affects the msvc2022* jobs,
and both these images have that same version of MSVC, but windows-2025
will be supported for much longer.

Replace windows-latest with windows-2025 for cygwin and msys2-* jobs.
The former is spewing warnings about changing from -2022 to -2025.
It's probably better to be explicit anyway.  We'll have to explicitly
update every few years, but that's not onerous.