CI: Drop macos-13 job
This macOS version is no longer supported by Apple or homebrew, and
our job has started to fail to install the required homebrew packages.
Github have deprecated the image and plan to remove it in early
December, so it doesn't seem worth directing any further effort into
keeping it going at this point.