Adjust type of unique_ptr<>
The object here is always a GlassTableCheck, but we only use it as
a GlassTable (its parent class).

The unique_ptr was holding a pointer to GlassTable, which should
work equally well (none of the methods of either class are virtual),
but valgrind complains, seemingly incorrectly - reported as:

https://bugs.kde.org/show_bug.cgi?id=511932

It's easy to workaround here by just changing the type so let's do
that so we can continue to test with current valgrind releases.

(cherry picked from commit 2790851691343815531a06dc502ee772ae8591e9)