Sync UTF-8 handling fix from Snowball
Fix a bug in Snowball's runtime code for decoding 4-byte UTF-8
sequences.  This affects grouping checks against Unicode codepoints
U+40000 to U+7FFFF and U+80000 to U+FFFFF, and we would decode a higher
codepoint value.  The largest codepoint used in a grouping in any
stemmer we ship is U+0585 so this seems to be a latent bug.
Reported in Snowball by Stephane Carrez.