Protect fds from stray writes on Microsoft Windows
We don't have the POSIX standard function pwrite(), so we need to
use a different technique.

We now open a database table file for writing using the low-level,
platform-specific function CreateFile() and specify
FILE_FLAG_OVERLAPPED, then wrap it as a file descriptor.  This gives
a file descriptor for which write() always fails with EINVAL, thus
protecting it from strays writes.

Then to write to the file descritor we extract the proprietary HANDLE
and use low-level function WriteFile() to write at a specified
position.

Tests show this is actually a little faster than using lseek() and
read().

We need to be able to read blocks from a file opened read-only or
read-write, so this means we also need to use ReadFile() to read
a block.  We don't specify FILE_FLAG_OVERLAPPED in this case as
it isn't needed (because the file descriptor can't be written to
anyway), and using it adds some overhead (particularly for reading
it seems - perhaps because for disk based IO writes always seem to
be handled synchronously whereas reads give ERROR_IO_PENDING and
we have to call GetOverlappedResult() to wait for them to complete).

Partly addresses #651