Refactor diversification API
Eliminate the separate class and instead have an MSet::diversify()
method which mutates the MSet object it is called on.

This is simpler and feels less clumsy to use.  The biggest gain is
that the diversified results are now accessible via an MSet object
just like normal results.

Partly addresses #804.