Avoid invalid buffer overwrite in `GlassTable::compact`
When a database is corrupted, there are possibly unexpectedly large
number or size of LeafItems or BItems, which may cause the `e` in
`GlassTable::compact` to be decreased a negative value.

When `e` becomes negative, the following `memcpy` call(s) will overwrite
bytes before the start of the dynamically allocated buffer, which in
turn corrupts the heap (causing undefined behavior).