Use more `AssertEqParanoid` in `ValueStreamDocument::do_get_value`
When `XAPIAN_ASSERTIONS` is defined but `XAPIAN_ASSERTIONS_PARANOID` is
not, the `doc` is possibly null in the assertion. As a result, the
assertion can cause more undefined behavior, which seemingly defeats the
intent of assertions.

In the master branch, all assertions are removed from
`valuestreamdocument.cc`. But perhaps we should still fix the release
branch to make `XAPIAN_ASSERTIONS` usable.