Update snowball compiler and runtime
We were using a 6-year old version of the Snowball compiler, with a
number of Xapian-specific patches.

We're now using the compiler and runtime from upstream git commit
94d5e77b3ccee55681ed740dad0c3377dbd72e3a.  The only differences from
upstream are that `utilities.c` is renamed to `.cc` and the following
line added to it:

    #define SNOWBALL_RUNTIME_THROW_EXCEPTIONS

Most of the patches we were carrying have either been merged upstream
in some form, or eliminated in favour of new functionality upstream.

The exception is the patch we were carrying which created an overlapping
pool of string constants.  The benefits of this were that it reduced the
size of the constant data segment, and reduced the number of runtime
dynamic relocations.

The relocations are from strings used in among tables, and I'm working
on a new way to implement among which is significantly faster and avoids
this problem.

The size saving is probably fairly modest.  It would reduce cache
pressure, but on the flip side it might make loading a string address
slightly slower.  I plan to profile to see if the former change is
acutally beneficial, and if so I'll rebase it onto current Snowball
and get it merged upstream.