Add condition to Query from iterator pair ctor template
Only enable if std::iterator_traits<I>::iterator_category is a type
since that's a requirement for the template to be useful because
we use that type in the constructor body.

This condition is already present on git master (added as part of
the conversion to using std::string_view in the API).